app/requirements.txt
aiosqlite==0.17.0
asyncpg==0.29.0
alembic==1.13.3
fastapi==0.78.0
Jinja2==3.1.4
jmespath==1.0.1
jsonformatter==0.3.2
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pydantic==1.10.18
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
SQLAlchemy==2.0.35
urllib3==1.26.20
uvicorn==0.17.6
watchgod==0.8.2
accessify==0.3.1
aiofiles==23.2.1
aiogoogle==5.7.0
aiohttp==3.9.3
aiosignal==1.3.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.3
attrs==21.4.0
bcrypt==3.2.2
cachetools==5.3.3
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
Faker==12.0.1
fastapi-users==10.0.4
fastapi-users-db-sqlalchemy==4.0.3
flake8==4.0.1
freezegun==1.2.1
frozenlist==1.4.1
google-auth==2.29.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
isort==5.13.2
makefun==1.13.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.2.2
multidict==6.0.5
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.23.4
pytest-freezegun==0.4.2
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tonyg-rfc3339==0.1
uvloop==0.19.0
websockets==10.3
yarl==1.9.4

